change the vault name
