+++ src/components/VaultRow/index.ts
@@ -1,1 +1,2 @@
+export * from "./VaultNameCell"
 export * from "./VaultRow"
